+++ benchmarks/common/Generators.hpp
@@ -10,5 +10,5 @@ namespace Algorithms::Benchmarks::Common
     std::vector<int> GenerateRandom(size_t n);
     std::vector<int> GenerateSorted(size_t n);
     std::vector<int> GenerateReverse(size_t n);
-    Math::Matrix<int> CreateRandomMatrix(size_t rows, size_t cols);
+    Math::Matrix<double> CreateRandomMatrix(size_t rows, size_t cols);
 }
